# openapi.yml
openapi: 3.1.1
info:
  title: Rekapundi
  description: Rekap your pundi.
  version: 0.1.0
...
paths:
...
  /summaries/generate/raw:
    post:
      tags: [summaries]
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/GenerateSummary"
      responses:
        "200":
          $ref: "#/components/responseBodies/SummaryRaw"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRelatedError"
        "401":
          description: Unauthorized.
        "422":
          description: Invalid request body semantics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRelatedError"
...
components:
  schemas:
...
    GenerateSummary:
      type: object
      properties:
        startDate:
          description: |
            The start date as the lower bound (inclusive) of the summary.
          type: string
          format: date
        endDate:
          description: |
            The start date as the higher bound (inclusive) of the summary.
          type: string
          format: date
        excludeCategoryIds:
          description: |
            The IDs of categories to exclude from the summary.
          type: array
          items:
            type: integer
      required: [startDate, endDate, excludeCategoryIds]
...
    Summary:
      description: The financial summary in JSON format.
      type: object
      properties:
        expense:
          type: object
          properties:
            amount:
              description: The total amount of expenses.
              type: integer
              minimum: 0
            groupSummary:
              description: |
                The summary of expenses grouped by categories and priorities.
                The categories are ordered by the amount in descending order.
              type: object
              properties:
                parentCategories:
                  description: |
                    The list of parent categories ordered by the amount in descending order.
                    Each parent category contains a list of child categories.
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        description: The name of the parent category.
                        type: string
                      amount:
                        description: The total amount of the parent category.
                        type: integer
                        minimum: 0
                      categories:
                        description: |
                          The list of child categories ordered by the amount in descending order.
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              description: The name of the child category.
                              type: string
                            amount:
                              description: The total amount of the child category.
                              type: integer
                              minimum: 0
                          required: [name, amount]
                    required: [amount, categories, name]
                priorities:
                  description: |
                    The list of priorities ordered by the amount in descending order.
                  type: array
                  items:
                    type: object
                    properties:
                      level:
                        description: |
                          The level of priority.
                          0 is the highest (primary needs), 1 is the secondary needs, 2 is the lowest priority (tertiary needs).
                        type: integer
                        minimum: 0
                        maximum: 2
                      amount:
                        description: The total amount of the priority.
                        type: integer
                        minimum: 0
                    required: [amount, level]
              required: [parentCategories, priorities]
          required: [amount, groupSummary]
        income:
          type: object
          properties:
            amount:
              description: The total amount of incomes.
              type: integer
              minimum: 0
            groupSummary:
              description: |
                The summary of incomes grouped by wallets.
                The wallets are ordered by the amount in descending order.
              type: object
              properties:
                wallets:
                  description: |
                    The list of wallets ordered by the amount in descending order.
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        description: The name of the wallet.
                        type: string
                        example: GoPay
                      amount:
                        description: The total amount of the wallet.
                        type: integer
                        minimum: 0
                    required: [amount, name]
              required: [wallets]
          required: [amount, groupSummary]
      required: [expense, income]
    ClientRelatedError:
      type: object
      properties:
        message:
          description: The error message that explains the invalid request.
          type: string
      required: message
  requestBodies:
...
    GenerateSummary:
      description: The input to generate an income/expense summary.
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenerateSummary"
...
  responseBodies:
...
    SummaryRaw:
      description: The financial summary in JSON.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Summary"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT